pyproject.toml
[tool.poetry]
name = "colour-streamlit-tm-30-18"
version = "0.1.0"
description = "Generates the \"ANSI/IES TM-30-18 Colour Rendition Report\" using Colour and Streamlit."
license = "BSD-3-Clause"
authors = [ "Colour Developers <colour-developers@colour-science.org>" ]
maintainers = [ "Colour Developers <colour-developers@colour-science.org>" ]
readme = 'README.rst'
repository = "https://github.com/colour-science/colour-streamlit-tm-30-18"
homepage = "https://www.colour-science.org/"
keywords = [
    "color",
    "color-science",
    "color-space",
    "color-spaces",
    "colorspace",
    "colorspaces",
    "colour",
    "colour-science",
    "colour-space",
    "colour-spaces",
    "colourspace",
    "colourspaces",
    "streamlit",
    "docker",
    "python"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development"
]

[tool.poetry.dependencies]
python = "^3.6"
colour-science = "^0.3.16"
matplotlib = "*"
streamlit = "*"

coverage = { version = "*", optional = true }  # Development dependency.
flake8 = { version = "*", optional = true }  # Development dependency.
invoke = { version = "*", optional = true }  # Development dependency.
nose = { version = "*", optional = true }  # Development dependency.
pre-commit = { version = "*", optional = true }  # Development dependency.
pytest = { version = "*", optional = true }  # Development dependency.
yapf = { version = "0.23", optional = true }  # Development dependency.

[tool.poetry.dev-dependencies]
coverage = "*"
flake8 = "*"
invoke = "*"
nose = "*"
pre-commit = "*"
pytest = "*"
yapf = "*"

[tool.poetry.extras]
development = [
    "coverage",
    "flake8",
    "invoke",
    "nose",
    "pre",
    "pytest",
    "yapf"
]

[build-system]
requires = [ "poetry>=0.12" ]
build-backend = "poetry.masonry.api"
